RELMGMT-1464 The Merge Conflicts Tab will display only if the Pull Request is open.  That should take care of 99% of our crashes stemming from folks clicking the tab after a branch has been merged/removed.  I can't imagine there is a huge market for folks clicking "Merge Conflicts" on PRs that are already merged, right?
As a backup, I smartened up the Exception Handling so if somehow someone does end up with an Open PR with a either a source or destination branch non-existent, a crash will be averted, and a message indicating each missing branch will be rendered instead of the "Oops Crash" message.  Bitbucket doesn't let you remove a branch with open PRs against it anyway so I think we are covered.